src/modules/github/PullRequestService.test.ts
import PullRequestService from "@/modules/github/PullRequestService";
import { DatabaseService } from "@/modules/db/SupDatabaseService";

describe("PullRequestService", () => {
    let prService: PullRequestService;
    let dbService: DatabaseService;

    beforeEach(() => {
        dbService = new DatabaseService();
        prService = new PullRequestService("owner", "repo", "branch");
    });

    it("should retrieve files from the database if commit SHA matches", async () => {
        // Mock the database service to return a known commit SHA
        jest.spyOn(dbService, "getFileDetails").mockResolvedValueOnce({
            commitHash: "knownCommitSHA",
        });

        const files = await prService.createPullRequest(/* parameters */);
        expect(files).toBeDefined();
        // Add more assertions based on expected behavior
    });

    it("should fetch files from GitHub if commit SHA does not match", async () => {
        // Mock the database service to return a different commit SHA
        jest.spyOn(dbService, "getFileDetails").mockResolvedValueOnce({
            commitHash: "differentCommitSHA",
        });

        const files = await prService.createPullRequest(/* parameters */);
        expect(files).toBeDefined();
        // Add more assertions based on expected behavior
    });

    it("should update the database with the current commit SHA after processing", async () => {
        // Mock the database update method
        const updateSpy = jest.spyOn(dbService, "saveFileDetails").mockResolvedValueOnce(undefined);

        await prService.createPullRequest(/* parameters */);
        expect(updateSpy).toHaveBeenCalled();
        // Add more assertions based on expected behavior
    });
});
